refactor completeFileResult
